goto_rw: store function name alongside instruction reference
We are working towards removing the "function" field from
goto_programt::instructionst::instructiont, and thus need to pass the identifier
of the function whenever it actually is required. goto_rw does not yet make use
of this, but will need to once dereferencing is updated to require a function
name.